✅ Allow some test class to be flaky
Unfortunately, the Instrumented Tests are failing a lot in the CI due to
the Compose + Gradle Managed Devices usage. Since I'm not able to
reproduce any of those failures locally, the silver tape solution now
is basically creating a rule to allow them to run again.
The extra setup is required due to how Compose default rules creates the
Activities.

By having a single rule and just applying it to the Tests, it's possible
to have a single source of truth. Then we can just remove it in the
future when Managed Devices are more stable.

For more info, please take a look:
https://proandroiddev.com/managing-flaky-tests-in-jetpack-compose-89c598590068
