+++ app/src/androidTest/java/com/escodro/alkaa/CategoryFlowTest.kt
@@ -1,74 +1,47 @@
 package com.escodro.alkaa
 
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
 import androidx.annotation.StringRes
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.test.SemanticsMatcher
 import androidx.compose.ui.test.hasSetTextAction
 import androidx.compose.ui.test.junit4.ComposeTestRule
-import androidx.compose.ui.test.junit4.createEmptyComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
 import androidx.compose.ui.test.performTextReplacement
-import androidx.test.core.app.ActivityScenario
 import androidx.test.platform.app.InstrumentationRegistry
-import com.adevinta.android.barista.rule.flaky.AllowFlaky
-import com.adevinta.android.barista.rule.flaky.FlakyTestRule
 import com.escodro.alkaa.navigation.NavGraph
 import com.escodro.category.presentation.semantics.ColorKey
 import com.escodro.designsystem.AlkaaTheme
 import com.escodro.local.provider.DaoProvider
-import com.escodro.test.DisableAnimationsRule
+import com.escodro.test.FlakyTest
 import kotlinx.coroutines.runBlocking
-import org.junit.After
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 import org.koin.test.KoinTest
 import org.koin.test.inject
 import com.escodro.category.R as CategoryR
 
-internal class CategoryFlowTest : KoinTest {
+internal class CategoryFlowTest : FlakyTest(), KoinTest {
 
     private val daoProvider: DaoProvider by inject()
-
-    @get:Rule
-    val composeTestRule = createEmptyComposeRule()
-
-    @get:Rule
-    val flakyRule = FlakyTestRule()
-
-    private lateinit var scenario: ActivityScenario<ComponentActivity>
-
-    @get:Rule
-    val disableAnimationsRule = DisableAnimationsRule()
 
     private val context = InstrumentationRegistry.getInstrumentation().targetContext
 
     @Before
     fun setup() {
-        scenario = ActivityScenario.launch(ComponentActivity::class.java)
         runBlocking {
             // Clean all existing categories
             daoProvider.getCategoryDao().cleanTable()
         }
 
-        scenario.onActivity { activity ->
-            activity.setContent {
-                AlkaaTheme {
-                    NavGraph()
-                }
+        setContent {
+            AlkaaTheme {
+                NavGraph()
             }
         }
         navigateToCategory()
-    }
-
-    @After
-    fun tearDown() {
-        scenario.close()
     }
 
     @Test
@@ -113,10 +86,7 @@ internal class CategoryFlowTest : KoinTest {
         }
     }
 
-    // Flaky test only on the CI, works fine locally
-    // Please see https://issuetracker.google.com/issues/230857082
     @Test
-    @AllowFlaky(attempts = 5)
     fun test_updateCategoryColor() {
         val name = "Anime"
         val color = Color(0xFFFFCA28)
