+++ app/src/androidTest/java/com/escodro/alkaa/HomeScreenTest.kt
@@ -1,6 +1,5 @@
 package com.escodro.alkaa
 
-import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
@@ -8,24 +7,17 @@ import androidx.test.platform.app.InstrumentationRegistry
 import com.escodro.alkaa.model.HomeSection
 import com.escodro.alkaa.navigation.NavGraph
 import com.escodro.designsystem.AlkaaTheme
-import com.escodro.test.DisableAnimationsRule
+import com.escodro.test.FlakyTest
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 
-internal class HomeScreenTest {
-
-    @get:Rule
-    val composeTestRule = createComposeRule()
-
-    @get:Rule
-    val disableAnimationsRule = DisableAnimationsRule()
+internal class HomeScreenTest : FlakyTest() {
 
     private val context = InstrumentationRegistry.getInstrumentation().targetContext
 
     @Before
     fun setup() {
-        composeTestRule.setContent {
+        setContent {
             AlkaaTheme {
                 NavGraph()
             }
