+++ app/src/androidTest/java/com/escodro/alkaa/NotificationFlowTest.kt
@@ -3,7 +3,6 @@ package com.escodro.alkaa
 import android.app.Notification
 import androidx.annotation.StringRes
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.test.platform.app.InstrumentationRegistry
@@ -15,12 +14,10 @@ import com.escodro.local.model.AlarmInterval
 import com.escodro.local.model.Task
 import com.escodro.local.provider.DaoProvider
 import com.escodro.task.R
-import com.escodro.test.DisableAnimationsRule
-import kotlinx.coroutines.ExperimentalCoroutinesApi
+import com.escodro.test.FlakyTest
 import kotlinx.coroutines.runBlocking
 import org.junit.After
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 import org.koin.test.KoinTest
 import org.koin.test.inject
@@ -29,18 +26,11 @@ import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 import com.escodro.alarm.R as AlarmR
 
-@OptIn(ExperimentalCoroutinesApi::class)
-internal class NotificationFlowTest : KoinTest {
+internal class NotificationFlowTest : FlakyTest(), KoinTest {
 
     private val daoProvider: DaoProvider by inject()
 
     private val scheduleAlarm: ScheduleAlarm by inject()
-
-    @get:Rule
-    val composeTestRule = createComposeRule()
-
-    @get:Rule
-    val disableAnimationsRule = DisableAnimationsRule()
 
     private val context = InstrumentationRegistry.getInstrumentation().targetContext
 
@@ -50,7 +40,7 @@ internal class NotificationFlowTest : KoinTest {
         runBlocking {
             daoProvider.getTaskDao().cleanTable()
         }
-        composeTestRule.setContent {
+        setContent {
             AlkaaTheme {
                 NavGraph()
             }
