+++ app/src/androidTest/java/com/escodro/alkaa/TaskFlowTest.kt
@@ -4,7 +4,6 @@ import androidx.annotation.StringRes
 import androidx.compose.ui.test.SemanticsMatcher
 import androidx.compose.ui.test.hasSetTextAction
 import androidx.compose.ui.test.junit4.ComposeTestRule
-import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
@@ -18,12 +17,11 @@ import com.escodro.designsystem.AlkaaTheme
 import com.escodro.local.model.Category
 import com.escodro.local.provider.DaoProvider
 import com.escodro.task.presentation.category.ChipNameKey
-import com.escodro.test.DisableAnimationsRule
 import com.escodro.test.Events
+import com.escodro.test.FlakyTest
 import com.escodro.test.assertIsChecked
 import kotlinx.coroutines.runBlocking
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 import org.koin.test.KoinTest
 import org.koin.test.inject
@@ -31,15 +29,9 @@ import org.koin.test.mock.declare
 import java.util.Calendar
 import com.escodro.task.R as TaskR
 
-internal class TaskFlowTest : KoinTest {
+internal class TaskFlowTest : FlakyTest(), KoinTest {
 
     private val daoProvider: DaoProvider by inject()
-
-    @get:Rule
-    val composeTestRule = createComposeRule()
-
-    @get:Rule
-    val disableAnimationsRule = DisableAnimationsRule()
 
     private val context = InstrumentationRegistry.getInstrumentation().targetContext
 
@@ -60,7 +52,7 @@ internal class TaskFlowTest : KoinTest {
         // Replace Debouncer with a Immediate Executor
         declare<CoroutineDebouncer> { CoroutinesDebouncerFake() }
 
-        composeTestRule.setContent {
+        setContent {
             AlkaaTheme {
                 NavGraph()
             }
