+++ app/src/androidTest/java/com/escodro/alkaa/TaskListFlowTest.kt
@@ -4,7 +4,6 @@ import androidx.annotation.StringRes
 import androidx.compose.ui.test.SemanticsMatcher
 import androidx.compose.ui.test.hasSetTextAction
 import androidx.compose.ui.test.junit4.ComposeTestRule
-import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
@@ -16,24 +15,17 @@ import com.escodro.designsystem.AlkaaTheme
 import com.escodro.local.model.Category
 import com.escodro.local.provider.DaoProvider
 import com.escodro.task.presentation.list.CheckboxNameKey
-import com.escodro.test.DisableAnimationsRule
+import com.escodro.test.FlakyTest
 import kotlinx.coroutines.runBlocking
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 import org.koin.test.KoinTest
 import org.koin.test.inject
 import com.escodro.task.R as TaskR
 
-internal class TaskListFlowTest : KoinTest {
+internal class TaskListFlowTest : FlakyTest(), KoinTest {
 
     private val daoProvider: DaoProvider by inject()
-
-    @get:Rule
-    val composeTestRule = createComposeRule()
-
-    @get:Rule
-    val disableAnimationsRule = DisableAnimationsRule()
 
     private val context = InstrumentationRegistry.getInstrumentation().targetContext
 
@@ -50,7 +42,7 @@ internal class TaskListFlowTest : KoinTest {
                 getCategoryDao().insertCategory(Category(name = "Work", color = "#519872"))
             }
         }
-        composeTestRule.setContent {
+        setContent {
             AlkaaTheme {
                 NavGraph()
             }
