+++ features/task/src/androidTest/java/com/escodro/task/AlarmPermissionFlowTest.kt
@@ -1,49 +1,24 @@
 package com.escodro.task
 
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.test.junit4.createEmptyComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
-import androidx.test.core.app.ActivityScenario
 import androidx.test.platform.app.InstrumentationRegistry
 import com.escodro.designsystem.AlkaaTheme
 import com.escodro.task.fake.PermissionStateFake
 import com.escodro.task.presentation.detail.alarm.AlarmSelectionContent
 import com.escodro.task.presentation.detail.alarm.AlarmSelectionState
-import com.escodro.test.DisableAnimationsRule
+import com.escodro.test.FlakyTest
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.PermissionState
 import com.google.accompanist.permissions.PermissionStatus
-import org.junit.After
 import org.junit.Assert
-import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 
 @OptIn(ExperimentalPermissionsApi::class)
-internal class AlarmPermissionFlowTest {
-
-    @get:Rule
-    val composeTestRule = createEmptyComposeRule()
-
-    @get:Rule
-    val disableAnimationsRule = DisableAnimationsRule()
+internal class AlarmPermissionFlowTest : FlakyTest() {
 
     private val context = InstrumentationRegistry.getInstrumentation().targetContext
-
-    private lateinit var scenario: ActivityScenario<ComponentActivity>
-
-    @Before
-    fun setup() {
-        scenario = ActivityScenario.launch(ComponentActivity::class.java)
-    }
-
-    @After
-    fun tearDown() {
-        scenario.close()
-    }
 
     @Test
     fun test_notificationPermissionGrantedDoesNotShowDialog() {
@@ -148,18 +123,16 @@ internal class AlarmPermissionFlowTest {
         permissionState: PermissionState,
         hasAlarmPermission: Boolean
     ) {
-        scenario.onActivity { activity ->
-            activity.setContent {
-                AlkaaTheme {
-                    AlarmSelectionContent(
-                        context = LocalContext.current,
-                        state = state,
-                        permissionState = permissionState,
-                        hasAlarmPermission = { hasAlarmPermission },
-                        onAlarmUpdate = {},
-                        onIntervalSelect = {}
-                    )
-                }
+        setContent {
+            AlkaaTheme {
+                AlarmSelectionContent(
+                    context = LocalContext.current,
+                    state = state,
+                    permissionState = permissionState,
+                    hasAlarmPermission = { hasAlarmPermission },
+                    onAlarmUpdate = {},
+                    onIntervalSelect = {}
+                )
             }
         }
     }
