+++ features/task/src/androidTest/java/com/escodro/task/AlarmSelectionTest.kt
@@ -1,47 +1,22 @@
 package com.escodro.task
 
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.createEmptyComposeRule
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
-import androidx.test.core.app.ActivityScenario
 import androidx.test.platform.app.InstrumentationRegistry
 import com.escodro.designsystem.AlkaaTheme
 import com.escodro.task.model.AlarmInterval
 import com.escodro.task.presentation.detail.alarm.AlarmSelection
-import com.escodro.test.DisableAnimationsRule
 import com.escodro.test.Events
-import org.junit.After
-import org.junit.Before
-import org.junit.Rule
+import com.escodro.test.FlakyTest
 import org.junit.Test
 import java.util.Calendar
 
-internal class AlarmSelectionTest {
-
-    @get:Rule
-    val composeTestRule = createEmptyComposeRule()
-
-    @get:Rule
-    val disableAnimationsRule = DisableAnimationsRule()
+internal class AlarmSelectionTest : FlakyTest() {
 
     private val context = InstrumentationRegistry.getInstrumentation().targetContext
-
-    private lateinit var scenario: ActivityScenario<ComponentActivity>
-
-    @Before
-    fun setup() {
-        scenario = ActivityScenario.launch(ComponentActivity::class.java)
-    }
-
-    @After
-    fun tearDown() {
-        scenario.close()
-    }
 
     @Test
     fun test_addAlarm() {
@@ -169,18 +144,16 @@ internal class AlarmSelectionTest {
         hasExactAlarmPermission: Boolean = true,
         shouldAskForNotificationPermission: Boolean = false
     ) {
-        scenario.onActivity { activity ->
-            activity.setContent {
-                AlkaaTheme {
-                    AlarmSelection(
-                        calendar = null,
-                        interval = AlarmInterval.NEVER,
-                        onAlarmUpdate = {},
-                        onIntervalSelect = {},
-                        hasAlarmPermission = { hasExactAlarmPermission },
-                        shouldCheckNotificationPermission = shouldAskForNotificationPermission
-                    )
-                }
+        setContent {
+            AlkaaTheme {
+                AlarmSelection(
+                    calendar = null,
+                    interval = AlarmInterval.NEVER,
+                    onAlarmUpdate = {},
+                    onIntervalSelect = {},
+                    hasAlarmPermission = { hasExactAlarmPermission },
+                    shouldCheckNotificationPermission = shouldAskForNotificationPermission
+                )
             }
         }
     }
@@ -189,18 +162,16 @@ internal class AlarmSelectionTest {
         calendar: Calendar,
         alarmInterval: AlarmInterval
     ) {
-        scenario.onActivity { activity ->
-            activity.setContent {
-                AlkaaTheme {
-                    AlarmSelection(
-                        calendar = calendar,
-                        interval = alarmInterval,
-                        onAlarmUpdate = {},
-                        onIntervalSelect = {},
-                        hasAlarmPermission = { true },
-                        shouldCheckNotificationPermission = false
-                    )
-                }
+        setContent {
+            AlkaaTheme {
+                AlarmSelection(
+                    calendar = calendar,
+                    interval = alarmInterval,
+                    onAlarmUpdate = {},
+                    onIntervalSelect = {},
+                    hasAlarmPermission = { true },
+                    shouldCheckNotificationPermission = false
+                )
             }
         }
     }
