+++ libraries/test/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     id("com.escodro.android-library")
+    id("com.escodro.android-compose")
 }
 
 android {
@@ -22,7 +23,8 @@ dependencies {
 
     api(libs.coroutines.test)
 
-    implementation(libs.compose.uitest) {
+    api(libs.compose.activity)
+    implementation(libs.bundles.composetest) {
         exclude(group = "androidx.core", module = "core-ktx")
         exclude(group = "androidx.fragment", module = "fragment")
         exclude(group = "androidx.customview", module = "customview")
